+++ changelog.md
@@ -6,6 +6,7 @@
 - Specified Python version 3.8.12 or higher.
 - Added link to transaction details on the token request response.
 - Added logging and error handling.
+- Added daily caps per faucet.
 
 ## v0.1